Improve pronunciation discovery
The app will now stop searching as soon as a suitable alternative is found. The motivation for this is that multiple occurrences of the same set of hiragana may be present in the search, but only one should include intonation specifics. Other occurrences, which disregard intonation, appear to be things like movie or song titles; furthermore, they tend to appear last in the list of alternatives.